Add files via upload
Code 1: Just Krill.
For varying iterations we see the unimpeded evolution of the Krill population.
Observe that the Krill 
Which shows that when Krill has no predator it’s population rapidly grows to its maximum capacity.
The value of k(n) approaches for a large value of n is 1. This is the equilibrium point of the population.
The equilibrium point denotes the carrying capacity of the habitation of the Krill, which is the size of the population of Krill which their environment can support.
Since k(n)=q(n)/Q and k(n) ϵ [0.1,1], then k(n) is a fraction denoting the existing Krill with respect to the carrying capacity of the Krill population.
If q(n) is the quantity of existing Krill 	which has easily reached the maximum level without the presence of predators, then q(n)=Q.





Where k(n) = q(n)/Q = Q/Q = 1
Substituting the k(n) = 1 into the recurrence equation:
k_{n+1}\ =k_n+r_1\ast k_n\left(1-k_n\right) 
	=\frac{Q}{Q}+r_1\ast\frac{Q}{Q}(1-\frac{Q}{Q})
 	 =1+r_1\ast1(1-1)
	= 1 +0
	= 1

The r1 coefficient is the ‘intrinsic growth rate’ which is the rate of population increase if there is no density-dependent regulation of the population.